test: move extraction goldens to query-backed seam
