+++ crates/djls-testing/src/lib.rs
@@ -1,5 +1,6 @@
 mod corpus;
 mod db;
+mod extraction;
 mod fixtures;
 mod mdtest;
 mod vendor;
@@ -15,6 +16,10 @@ pub use corpus::lock_corpus;
 pub use corpus::module_path_from_file;
 pub use corpus::sync_corpus;
 pub use db::TestDatabase;
+pub use extraction::ExtractionBundle;
+pub use extraction::SortedExtractionResult;
+pub use extraction::extract_bundle;
+pub use extraction::sorted_snapshot;
 pub use fixtures::ProjectFixture;
 pub use fixtures::build_entry_specs;
 pub use fixtures::build_specs_from_extraction;
